+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added mysql db for keyrock to terraform
+- Added keyrock and wilma integration
 - Use efs instead of ebs volume in wirecloud
 - Added nginx to wirecloud deployment
 - Fixed ingress path value typo